Implemented Services

+++ src/main/java/com/github/SampleRESTApi/services/StudentServices.java
@@ -1,46 +1,51 @@
-package com.github.SampleRESTApi.controllers;
+package com.github.SampleRESTApi.services;
 
 import com.github.SampleRESTApi.models.Student;
 import com.github.SampleRESTApi.repository.StudentRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
-@RestController
-public class StudentsController {
+public class StudentServices {
+    private final StudentRepository studentRepository;
 
-    @Autowired
-    private StudentRepository studentRepository;
+    public StudentServices(StudentRepository studentRepository) {
+        this.studentRepository = studentRepository;
+    }
 
-    @ResponseBody
-    @GetMapping(value = "/students")
     public List<Student> getStudents(){
         return studentRepository.findAll();
     }
 
-    @PostMapping("/students")
-    public Student saveStudent(@RequestBody Student student){
+    public Student getStudentById(String id){
+        return studentRepository.findById(id).get();
+    }
+
+    public Student save(Student student){
         return studentRepository.save(student);
     }
 
-    @PutMapping("/students/{id}")
-    public Student updateUser(@PathVariable String id, @RequestBody Student student){
+    public Student update(String id, Student student){
         Student updatedStudent = studentRepository.findById(id).get();
+        if(updatedStudent == null){
+            return null;
+        }
         updatedStudent.setFirstName(student.getFirstName());
         updatedStudent.setMiddleName(student.getMiddleName());
         updatedStudent.setLastName(student.getLastName());
         updatedStudent.setAge(student.getAge());
         updatedStudent.setYearLevel(student.getYearLevel());
         updatedStudent.setSection(student.getSection());
-
-        return studentRepository.save(updatedStudent);
+        studentRepository.save(updatedStudent);
+        return updatedStudent;
     }
 
-    @DeleteMapping("/students/{id}")
-    public String updateUser(@PathVariable String id){
+    public Student deleteById(String id){
         Student deleteStudent = studentRepository.findById(id).get();
+        if (deleteStudent == null) {
+            return null;
+        }
         studentRepository.delete(deleteStudent);
-        return "Deleted " + id + "...";
+        return deleteStudent;
     }
 }
